rename, remove logs, added comments
