+++ src/vim-webgl-viewer/viewer.ts
@@ -11,7 +11,7 @@ import { Input } from './input'
 import { Selection } from './selection'
 import { Environment, IEnvironment } from './environment'
 import { Raycaster, RaycastResult } from './raycaster'
-import { CameraGizmo } from './gizmos'
+import { CameraGizmo } from './gizmoOrbit'
 import { RenderScene } from './renderScene'
 import { Viewport } from './viewport'
 import { GizmoAxes } from './gizmoAxes'
